abstract all the calendar stuff to a util and display time zones in layout publishing
